backend(user-query): move getSuggestion query to server action
