+++ actions/user.actions.ts
@@ -2,7 +2,7 @@
 import { db } from '@/lib/db/connect_db';
 import { users } from '@/lib/db/schema/schema';
 import { CreateUser, UpdateUser, User } from '@/types/db.types';
-import { eq } from 'drizzle-orm';
+import { and, eq, ilike, notInArray } from 'drizzle-orm';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 import 'server-only';
@@ -51,5 +51,44 @@ export async function deleteUser(clerkId: User['clerkId']): Promise<void> {
     redirect('/sign-up');
   } catch (error) {
     console.error(error);
+  }
+}
+
+export async function getSuggestedUsersByEmail(
+  input: string | null,
+  filterIds: string[] = [],
+): Promise<User[] | []> {
+  try {
+    console.log(filterIds);
+
+    // get suggested users by chunks of eight
+    const suggestedUsers = await db
+      .select()
+      .from(users)
+      .where(
+        and(
+          input ? ilike(users.primaryEmailAddress, `%${input.trim()}%`) : undefined,
+          filterIds.length > 0 ? notInArray(users.id, filterIds) : undefined,
+        ),
+      )
+      .limit(8);
+
+    console.log(filterIds);
+    console.log(suggestedUsers);
+
+    return suggestedUsers;
+    // return {
+    //   success: true,
+    //   message: 'Successfully retrieve all users',
+    //   data: suggestedUsers,
+    // };
+  } catch (error) {
+    console.error(error);
+    // return {
+    //   success: true,
+    //   message: 'Failed to retrieved users',
+    //   error: JSON.stringify(error),
+    // };
+    return [];
   }
 }
