+++ components/theme/theme-toggle.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { Moon, Sun } from 'lucide-react';
+import { useTheme } from 'next-themes';
+
+export function ThemeToggle() {
+  const { theme, setTheme } = useTheme();
+
+  return (
+    <button
+      onClick={() => setTheme((prev) => (prev === 'light' ? 'dark' : 'light'))}
+      className="bg-bg-primary rounded-4xl p-2 hover:cursor-pointer"
+      aria-label="Toggle theme"
+    >
+      {theme === 'light' ? <Moon size={20} /> : <Sun size={20} />}
+    </button>
+  );
+}
